644 feature abstract away identifier and expose a new parameter name (#646)
## What does this add?

We have changed the interface for `Session` so that instead of the
`identifier` param it takes in a `name` parameter. (note that for the
sake of this PR I just renamed the symbol [this worked because we were
using the parameter as a name anyway]).

The other major change is the json schema outputted. Instead of sending
a flattened list, we instead send top level with session _id,
session_name, start_time, end_time with an property "runs" which
contains the list of runs as before.
## Type of changes

Please check the type of change your PR introduces:

- [ ] 🐛 Bug fix (non-breaking change which fixes an issue)
- [x] ✨ New feature (non-breaking change which adds functionality)
- [x] 💥 Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 📚 Documentation update (improvements or corrections to
documentation)
- [ ] 🎨 Code style/formatting (changes that do not affect the meaning of
the code)
- [ ] ♻️ Refactoring (code change that neither fixes a bug nor adds a
feature)
- [ ] ⚡ Performance improvement (code change that improves performance)
- [ ] ✅ Test update (adding missing tests or correcting existing tests)
- [ ] 🔧 Build/CI changes (changes to build process or continuous
integration)
- [ ] 🗑️ Chore (other changes that don't modify src or test files)

## Background context

This was a needed adjustment because we needed to differentiate between
identifier and name. Our logic didn't make a ton of sense becuase we
were using the identifier more as a name than an identifier making the
lives of anyone working with `json` object difficult.


## Checklist for Author
<!-- Skip sections not relevant to your change type (e.g., skip Testing
for docs-only changes) -->

### Code Quality
- [x] Code follows the project's style guidelines (run `ruff check .`
and `ruff format .`)
- [x] Code is commented, particularly in hard-to-understand areas

### Testing
- [x] Tests added/updated and pass locally (`pytest tests`)
- [x] Test coverage maintained

### Documentation
- [x] Documentation updated if needed (bot will verify)

### Git & PR Management
- [x] PR title clearly describes the change