Rewrite and refactor unit test for nodes and built_nodes folder (#698)
## What does this add?
Write tests with near full line coverage of files that should/can be
tested in a unit test setting for both nodes and built_nodes. Note that
there are certain files that do not have units tests associated with
them, this is due to the OOP structure of our files and many factory
functions used. It makes it so many files in easy usage are wrapped up
and cannot be tested in a unit test setting. These are left for
integration testing.

## Type of changes

Please check the type of change your PR introduces:

- [ ] 🐛 Bug fix (non-breaking change which fixes an issue)
- [ ] ✨ New feature (non-breaking change which adds functionality)
- [ ] 💥 Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 📚 Documentation update (improvements or corrections to
documentation)
- [ ] 🎨 Code style/formatting (changes that do not affect the meaning of
the code)
- [x] ♻️ Refactoring (code change that neither fixes a bug nor adds a
feature)
- [ ] ⚡ Performance improvement (code change that improves performance)
- [x] ✅ Test update (adding missing tests or correcting existing tests)
- [ ] 🔧 Build/CI changes (changes to build process or continuous
integration)
- [ ] 🗑️ Chore (other changes that don't modify src or test files)

## Background context
Due to recent changes in overall structure of our repo testing had to be
updated.

## Checklist for Author
<!-- Skip sections not relevant to your change type (e.g., skip Testing
for docs-only changes) -->

### Code Quality
- [x] Code follows the project's style guidelines (run `ruff check .`
and `ruff format .`)
- [x] Code is commented, particularly in hard-to-understand areas

### Testing
- [x] Tests added/updated and pass locally (`pytest tests`)
- [x] Test coverage maintained

### Git & PR Management
- [x] PR title clearly describes the change